+++ src/main/java/com/kustaurant/kustaurant/evaluation/review/ReviewWebController.java
@@ -21,7 +21,7 @@ public class ReviewWebController {
 
     @GetMapping("/web/api/restaurants/{restaurantId}/comments")
     public ResponseEntity<List<ReviewsResponse>> getReviewList(
-            @PathVariable Integer restaurantId,
+            @PathVariable Long restaurantId,
             @RequestParam(defaultValue = "POPULARITY") @Parameter(description = "인기순: POPULARITY, 최신순: LATEST") SortOption sort,
             @Parameter(hidden = true) @AuthUser AuthUserInfo user
     ) {
