+++ src/main/java/com/kustaurant/kustaurant/post/post/infrastructure/jpa/PostPhotoJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface PostPhotoJpaRepository extends JpaRepository<PostPhotoEntity, Integer> {
-    List<PostPhotoEntity> findByPostId(Integer postId);
-    void deleteByPostId(Integer postId);
+    List<PostPhotoEntity> findByPostId(Long postId);
+    void deleteByPostId(Long postId);
 }
